+++ src/views/proj/SideBar/WorkspaceItem.vue
